# swagger/yaml/swagger-workflow.yaml
swagger: "2.0"
info: 
   title: "API for VMS Workflow Managment"
   description: "VMS Workflow Management APIs"
   version: 1.0.0
basePath: /api/v1
paths:
  /processDefinition/deploy/upload:
    post:
      tags:
      - "bpmn process definition managment"
      summary: "上传 bpmn zip 文件部署"
      description: ""
      operationId: uploadZip
      consumes:
         - multipart/form-data
      produces:
      - application/json
      parameters:
      - name: file
        in: formData
        type: file
        description: The file to upload.
        required: true
      - name: bpmnName
        in: query
        required: false
        type: string
      - name: category
        in: query
        required: false
        type: string
      responses:
        200:
          description: upload bpmn file  successfully
          schema:
            $ref: "#/definitions/StandardResponse"
  /processDefinition/deploy/predefined:
    post:
      tags:
      - "bpmn process definition managment"
      summary: "部署预定义的bpmn文件"
      description: ""
      operationId: deployDefinedProcess
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: bpmnName
        in: query
        required: true
        type: string
      - name: category
        in: query
        required: false
        type: string
      responses:
        200:
          description: upload bpmn file  successfully
          schema:
            $ref: "#/definitions/StandardResponse"          
...
  /processDefinition/deploy/{id}:
    get:
      tags:
      - "bpmn process definition managment"
      summary: "根据deployment ID查询详情"
      description: ""
      operationId: getProcessById
      produces:
        - application/json
      parameters: 
        - name: id
          required: true
          in: path
          type: integer
      responses: 
        200:
          description: "get deployment by id"
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
              data:
                $ref: "#/definitions/DeploymentEntity"
    delete:
      tags:
      - "bpmn process definition managment"
      summary: "根据deployment ID删除"
      description: ""
      operationId: deleteProcessById
      produces:
        - application/json
      parameters: 
        - name: id
          required: true
          in: path
          type: integer
      responses: 
        200:
          description: "get deployment by id"
          schema:
            $ref: "#/definitions/StandardResponse"
...
definitions: 
  DeploymentEntity:
    properties:
      id:
        type: integer
      name:
        type: string
      category:
        type: string
      key:
        type: string
      tenantId:
        type: string
      engineVersion:
        type: string
...
  StandardResponse:
    required: 
       - code
       - message
       - data
    properties:
      code:
        type: integer
      message:
        type: string
      data:
        type: object
...